requirements.txt
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.2.0
customtkinter>=5.0.0
matplotlib>=3.4.0
numpy>=1.20.0
streamlit>=1.22.0
opencv-python>=4.5.0
requests>=2.25.1 